+++ test_script.php
@@ -1,0 +1,3 @@
+<?php
+$history = new stdClass();
+$items = isset($history['items']) ? $history['items'] : array();