+++ test_script4.php
@@ -1,0 +1,4 @@
+<?php
+$history = new stdClass();
+$items = (is_array($history) && isset($history['items'])) ? $history['items'] : array();
+var_dump($items);